requirements.txt
joblib==1.2.0
mmcv==1.4.5
numpy==1.24.3
pandas==1.5.3
PyYAML==6.0.1
rdkit==2022.9.1
rdkit_pypi==2021.9.4
scikit_learn==1.2.2
tqdm==4.65.0
